+++ src/css/mod.rs
@@ -1,6 +1,7 @@
 //! Types for css values and rules.
 mod call_args;
 mod comment;
+mod item;
 mod rule;
 mod selectors;
 mod string;
@@ -10,7 +11,8 @@ mod valueformat;
 
 pub use self::call_args::CallArgs;
 pub use self::comment::Comment;
-pub use self::rule::{BodyItem, Import, Rule};
+pub use self::item::{Import, Item};
+pub use self::rule::{BodyItem, Property, Rule};
 pub use self::selectors::{BadSelector, Selector, SelectorPart, Selectors};
 pub use self::string::CssString;
 pub use self::value::{Value, ValueMap, ValueToMapError};