+++ tests/spec/css/plain/error/expression/variable.rs
@@ -8,7 +8,7 @@ fn runner() -> crate::TestRunner {
 }
 
 #[test]
-#[ignore] // missing error
+#[ignore] // wrong error
 fn declaration() {
     let runner = runner().with_cwd("declaration");
     assert_eq!(
@@ -23,7 +23,7 @@ fn declaration() {
     );
 }
 #[test]
-#[ignore] // missing error
+#[ignore] // wrong error
 fn test_use() {
     let runner = runner().with_cwd("use");
     assert_eq!(