+++ tests/spec/directives/test_use/error/load.rs
@@ -327,7 +327,7 @@ fn missing() {
     );
 }
 #[test]
-#[ignore] // missing error
+#[ignore] // wrong error
 fn no_extension() {
     let runner = runner().with_cwd("no_extension");
     assert_eq!(